Major changes
This commit includes several changes, including:
- Moved to QEMU 2.1.0
- Support for 64-bit guests
- Support for Linux and OS X targets
- Support for external plugins
- Refactoring & optimizations

Plus several other changes.

+++ src/roms/seabios/src/byteorder.h
@@ -1,6 +1,8 @@
 #ifndef __BYTEORDER_H
 #define __BYTEORDER_H
 
+#include "types.h" // u32
+
 static inline u16 __swab16_constant(u16 val) {
     return (val<<8) | (val>>8);
 }
@@ -43,7 +45,7 @@ static inline u16 le16_to_cpu(u16 x) {
 static inline u32 le32_to_cpu(u32 x) {
     return x;
 }
-static inline u32 le64_to_cpu(u64 x) {
+static inline u64 le64_to_cpu(u64 x) {
     return x;
 }
 
@@ -62,7 +64,7 @@ static inline u16 be16_to_cpu(u16 x) {
 static inline u32 be32_to_cpu(u32 x) {
     return swab32(x);
 }
-static inline u32 be64_to_cpu(u64 x) {
+static inline u64 be64_to_cpu(u64 x) {
     return swab64(x);
 }
